№6 v1.0 Перенос констант
